PINOT-1478: Adding implementation for ReaderWriter for SingleColumnSingleValue and SingleColumnMultiValue
RB=470974
R=xiafu,jfim,dpatel
A=jfim

+++ pinot-core/src/main/java/com/linkedin/pinot/core/index/writer/SingleColumnSingleValueWriter.java
@@ -42,7 +42,7 @@ public interface SingleColumnSingleValueWriter extends DataFileWriter {
    * @param row
    * @param l
    */
-  public void setLong(int row, int l);
+  public void setLong(int row, long l);
 
   /**
    *
