+++ pinot-core/src/main/java/com/linkedin/pinot/core/index/writer/impl/DefaultSingleColumnSingleValueWriter.java
@@ -57,7 +57,7 @@ public class DefaultSingleColumnSingleValueWriter implements
   }
 
   @Override
-  public void setLong(int row, int l) {
+  public void setLong(int row, long l) {
     dataFileWriter.setLong(row, 0, l);
 
   }
